+++ tests/TestModel.php
@@ -2,6 +2,7 @@
 
 namespace Spatie\MediaLibrary\Test;
 
+use Spatie\MediaLibrary\Media;
 use Illuminate\Database\Eloquent\Model;
 use Spatie\MediaLibrary\HasMedia\HasMediaTrait;
 use Spatie\MediaLibrary\HasMedia\Interfaces\HasMediaConversions;
@@ -19,7 +20,7 @@ class TestModel extends Model implements HasMediaConversions
      *
      * @return array
      */
-    public function registerMediaConversions()
+    public function registerMediaConversions(Media $media = null)
     {
     }
 }